Move codemod to index.js
